spark-python/pyspark-pyarrow/pyproject.toml
[tool.poetry]
name = "pyspark-pyarrow"
version = "0.1.0"
description = ""
authors = ["mahfooz_iiitian <mahfooz.iiitian@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyspark = {extras = ["sql"], version = "^3.5.1"}
ipykernel = "^6.29.5"


[tool.poetry.group.dev.dependencies]
pytest-mock = "^3.14.0"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
pythonpath = ["src"]
testpaths = [
    "tests",
    "integration",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
